requirements.txt
langchain-core
langchain-groq
python-dotenv
streamlit
python-docx
PyPDF2